# feat: Complete Phase 3 STT+Sentiment Integration & Performance Optimization