… apps/server/src/__tests__/assemblyai-integration.test.ts
… apps/server/src/services/optimizedCoachingService.ts
… docs/postman/API-Testing-Guide.md